Finished writing sign up and login, ready for testing
