+++ app/src/main/java/com/cmput301/w19t06/theundesirablejackals/Database/DatabaseHelper.java
@@ -2,12 +2,11 @@ package com.cmput301.w19t06.theundesirablejackals.Database;
 
 import android.content.Context;
 import android.content.Intent;
-import android.provider.ContactsContract;
 import android.support.annotation.NonNull;
 import android.widget.Toast;
 
-import com.cmput301.w19t06.theundesirablejackals.MainActivity;
-import com.cmput301.w19t06.theundesirablejackals.RegisterActivity;
+import com.cmput301.w19t06.theundesirablejackals.LoginActivity;
+import com.cmput301.w19t06.theundesirablejackals.MainHomeViewActivity;
 import com.cmput301.w19t06.theundesirablejackals.User.User;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -18,8 +17,6 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.HashMap;
-
-import static com.firebase.ui.auth.AuthUI.getApplicationContext;
 
 public class DatabaseHelper {
     // database path to users
@@ -59,11 +56,12 @@ public class DatabaseHelper {
                             HashMap<String, User> hashMap = new HashMap<>();
                             hashMap.put(username, new User(username, email, phonenumber));
 
+                            // save class user to database
                             databaseReference.setValue(hashMap).addOnCompleteListener(new OnCompleteListener<Void>() {
                                 @Override
                                 public void onComplete(@NonNull Task<Void> task) {
                                     if (task.isSuccessful()) {
-                                        Intent intent = new Intent(context, MainActivity.class);
+                                        Intent intent = new Intent(context, MainHomeViewActivity.class);
                                         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
                                         context.startActivity(intent);
                                     }
@@ -72,6 +70,22 @@ public class DatabaseHelper {
                         } else {
                             Toast.makeText(context,
                                     "You can't register with this email and or password", Toast.LENGTH_SHORT);
+                        }
+                    }
+                });
+    }
+
+    public void login(String email, String password) {
+        firebaseAuth.signInWithEmailAndPassword(email, password)
+                .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if(task.isSuccessful()) {
+                            Intent intent = new Intent(context, MainHomeViewActivity.class);
+                            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+                            context.startActivity(intent);
+                        } else {
+                            Toast.makeText(context, "Authentication Failed", Toast.LENGTH_SHORT);
                         }
                     }
                 });
