--- app/src/main/java/com/cmput301/w19t06/theundesirablejackals/RegisterActivity.java
@@ -1,108 +1,0 @@
-package com.cmput301.w19t06.theundesirablejackals;
-
-import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.text.TextUtils;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
-
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.rengwuxian.materialedittext.MaterialEditText;
-
-import java.util.HashMap;
-
-public class RegisterActivity extends AppCompatActivity {
-    EditText username, email, phone, password;
-    Button btn_register;
-
-    FirebaseAuth auth;
-    DatabaseReference reference;
-
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
-
-        username = findViewById(R.id.new_username);
-        email = findViewById(R.id.new_email);
-        phone = findViewById(R.id.new_phone);
-        password = findViewById(R.id.new_password);
-
-        btn_register = findViewById(R.id.btn_register);
-
-        auth = FirebaseAuth.getInstance();
-
-        btn_register.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String txt_username = username.getText().toString();
-                String txt_email = email.getText().toString();
-                String txt_password = password.getText().toString();
-                String txt_phone = phone.getText().toString();
-
-                if (TextUtils.isEmpty(txt_username) || TextUtils.isEmpty(txt_email)
-                    || TextUtils.isEmpty(txt_phone) || TextUtils.isEmpty(txt_password)) {
-
-                    Toast.makeText(RegisterActivity.this, "Empty field must be filled", Toast.LENGTH_SHORT);
-
-                } else if (txt_password.length() < 8) {
-                    Toast.makeText(RegisterActivity.this, "Password must be 8 characters or more", Toast.LENGTH_SHORT);
-                } else {
-                    register(txt_username, txt_email, txt_password);
-                }
-
-            }
-        });
-    }
-
-    private void register(final String username, String email, String password) {
-
-        auth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            FirebaseUser firebaseUser = auth.getCurrentUser();
-
-                            assert firebaseUser != null;
-                            String userid = firebaseUser.getUid();
-
-                            reference = FirebaseDatabase.getInstance().getReference("users").child(userid);
-
-                            HashMap<String, String> hashMap = new HashMap<>();
-                            hashMap.put("id", userid);
-                            hashMap.put("username", username);
-                            hashMap.put("imageURL", "default");
-
-                            reference.setValue(hashMap).addOnCompleteListener(new OnCompleteListener<Void>() {
-                                @Override
-                                public void onComplete(@NonNull Task<Void> task) {
-                                    if (task.isSuccessful()) {
-                                        Intent intent = new Intent(RegisterActivity.this, MainActivity.class);
-                                        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                                        startActivity(intent);
-                                        finish();
-                                    }
-                                }
-                            });
-                        } else {
-                            Toast.makeText(RegisterActivity.this,
-                                    "You can't register with this email and or password", Toast.LENGTH_SHORT);
-                        }
-                    }
-                });
-    }
-}
