+++ app/src/main/res/layout/activity_login.xml
@@ -5,65 +5,64 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".LoginActivity">
 
     <EditText
-        android:id="@+id/username"
+        android:id="@+id/etext_login_email"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_marginStart="104dp"
         android:layout_marginTop="179dp"
-        android:drawableLeft="@drawable/ic_action_username"
         android:ems="10"
-        android:hint="@string/user_name"
-        android:inputType="textPersonName" />
+        android:hint="@string/fui_email_hint"
+        android:inputType="textEmailAddress"/>
 
     <EditText
-        android:id="@+id/editText2"
+        android:id="@+id/etext_login_password"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignEnd="@+id/username"
+        android:layout_alignEnd="@+id/etext_login_email"
         android:layout_alignParentTop="true"
         android:layout_marginTop="266dp"
         android:layout_marginEnd="-1dp"
-        android:drawableStart="@drawable/ic_action_pass"
         android:ems="10"
-        android:hint="@string/password"
-        android:inputType="textPassword" />
+        android:hint="@string/fui_password_hint"
+        android:inputType="textPassword"/>
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
         <Button
-            android:id="@+id/buttonLogin"
+            android:id="@+id/btn_login"
             android:layout_width="255dp"
             android:layout_height="68dp"
             android:layout_marginStart="8dp"
             android:layout_marginTop="128dp"
             android:layout_marginEnd="8dp"
             android:background="@color/colorAccent"
-            android:onClick="loginBtn"
-            android:text="Login"
-            android:textColor="#fff"
+            android:text="@string/login"
+            android:textColor="@color/white"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/signupInitial_id" />
+            app:layout_constraintTop_toBottomOf="@+id/btn_login_signup" />
 
         <Button
-            android:id="@+id/signupInitial_id"
+            android:id="@+id/btn_login_signup"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="8dp"
+            android:layout_marginTop="341dp"
             android:layout_marginEnd="8dp"
-            android:onClick="signupBtn"
-            android:text="Sign Up"
+            android:layout_marginBottom="128dp"
+            android:text="@string/btn_signup_text"
+            app:layout_constraintBottom_toTopOf="@+id/btn_login"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.498"
             app:layout_constraintStart_toStartOf="parent"
-            tools:layout_editor_absoluteY="342dp" />
+            app:layout_constraintTop_toTopOf="parent" />
     </android.support.constraint.ConstraintLayout>
