+++ app/src/main/res/layout/activity_main_home_view.xml
@@ -1,76 +1,33 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.cmput301.w19t06.theundesirablejackals.MainHomeViewActivity"
-
-    android:orientation="vertical">
+    tools:context=".MainHomeViewActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:layout_gravity="center"
-        android:elevation="10dp"
-        android:background="#fff"
-        android:id="@+id/appbar_id">
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay">
 
-        <RelativeLayout
+        <android.support.v7.widget.Toolbar
+            android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <SearchView
-                android:layout_width="304dp"
-                android:layout_height="56dp"
-                android:layout_alignParentStart="true"
-                android:layout_alignParentTop="true"
-                android:layout_marginStart="17dp"
-                android:layout_marginTop="29dp"
-                android:background="@drawable/border"
-                android:orientation="vertical"
-                android:paddingTop="20dp"
-                tools:layout_editor_absoluteX="8dp"
-                tools:layout_editor_absoluteY="116dp">
-
-            </SearchView>
-
-            <ImageView
-                android:id="@+id/imageView"
-                android:layout_width="89dp"
-                android:layout_height="78dp"
-                android:layout_alignParentTop="true"
-                android:layout_alignParentEnd="true"
-                android:layout_marginStart="-17dp"
-                android:layout_marginTop="23dp"
-                android:layout_marginEnd="0dp"
-                android:layout_marginBottom="8dp"
-                android:background="@drawable/ic_action_menu"
-                tools:layout_editor_absoluteX="329dp"
-                tools:layout_editor_absoluteY="166dp" />
-
-        </RelativeLayout>
+            android:layout_height="?attr/actionBarSize"
+            android:background="?attr/colorPrimary"
+            app:popupTheme="@style/AppTheme.PopupOverlay" />
 
     </android.support.design.widget.AppBarLayout>
 
-    <android.support.design.widget.TabLayout
-        android:layout_width="match_parent"
-        android:layout_height="80dp"
-        app:tabGravity="fill"
-        app:tabMode="fixed"
-        android:id="@+id/tablayout_id"
-        app:tabTextColor="@color/tabTextColor"
-        app:tabIndicatorColor="@color/tabIndicatorColor"
-        app:tabIndicatorHeight="5dp"
-        app:tabPaddingStart="12dp"
-        app:tabPaddingEnd="12dp"
-        app:tabBackground="@color/tabBackgroundColor"/>
-    <android.support.v4.view.ViewPager
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/viewpager_id"
-        ></android.support.v4.view.ViewPager>
+    <include layout="@layout/content_main" />
 
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom|end"
+        android:layout_margin="@dimen/fab_margin"
+        app:srcCompat="@android:drawable/ic_input_add" />
 
-
-</android.support.v7.widget.LinearLayoutCompat>
+</android.support.design.widget.CoordinatorLayout>
