--- app/src/main/res/layout/activity_register.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".RegisterActivity">
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:gravity="center_horizontal"
-        android:padding="16dp">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/tv_create_new_account"
-            android:textSize="16dp"
-            android:textStyle="bold"/>
-
-
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/new_username"
-            android:layout_marginTop="10dp"
-            app:met_floatingLabel="normal"
-            android:hint="@string/hint_username"/>
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/new_email"
-            android:layout_marginTop="10dp"
-            app:met_floatingLabel="normal"
-            android:hint="@string/hint_email"/>
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/new_phone"
-            android:layout_marginTop="10dp"
-            app:met_floatingLabel="normal"
-            android:hint="@string/hint_phone"/>
-
-        <com.rengwuxian.materialedittext.MaterialEditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/new_password"
-            android:layout_marginTop="10dp"
-            app:met_floatingLabel="normal"
-            android:hint="@string/hint_password"/>
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/btn_register"
-            android:text="Register"/>
-
-
-
-
-
-    </LinearLayout>
-
-
-
-
-
-</android.support.design.widget.CoordinatorLayout>
