+++ app/src/main/res/layout/activity_signup.xml
@@ -86,8 +86,7 @@
         android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
         android:layout_marginBottom="8dp"
-        android:text="@string/btn_signup_signup_text"
-        android:onClick="OnClickSignupButton"
+        android:text="@string/btn_signup_text"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
