+++ app/src/main/res/layout/content_main.xml
@@ -4,6 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:context=".MainHomeViewActivity"
+    tools:showIn="@layout/activity_main_home_view">
 
 </android.support.constraint.ConstraintLayout>
