+++ app/src/main/res/values/colors.xml
@@ -6,4 +6,5 @@
     <color name="tabTextColor">#10100E</color>
     <color name="tabIndicatorColor">#056fff</color>
     <color name="tabBackgroundColor">#a39398</color>
+    <color name="white">#fff</color>
 </resources>
