+++ app/src/main/res/values/strings.xml
@@ -8,5 +8,19 @@
     <string name="hint_email">Email</string>
     <string name="hint_phone">Phone number</string>
     <string name="tv_signup_create_new">Create a new account</string>
-    <string name="btn_signup_signup_text">Sign up</string>
+    <string name="btn_signup_text">Sign up</string>
+    <string name="title_activity_main">MainActivity</string>
+    <string name="title_activity_login3">Sign in</string>
+    <!-- Strings related to login -->
+    <string name="prompt_email">Email</string>
+    <string name="prompt_password">Password (optional)</string>
+    <string name="action_sign_in">Sign in or register</string>
+    <string name="action_sign_in_short">Sign in</string>
+    <string name="error_invalid_email">This email address is invalid</string>
+    <string name="error_invalid_password">This password is too short</string>
+    <string name="error_incorrect_password">This password is incorrect</string>
+    <string name="error_field_required">This field is required</string>
+    <string name="permission_rationale">"Contacts permissions are needed for providing email
+        completions."
+    </string>
 </resources>
